Split admin support-ticket storage into store_tickets.go
